fix(Submission) DTO,Controller,Entity,Mapper

+++ backend/src/main/java/com/example/starter_project_2025/system/assessment/dto/submission/response/SubmissionResultResponse.java
@@ -10,8 +10,8 @@ import java.util.UUID;
 public class SubmissionResultResponse {
 
     private UUID submissionId;
-    private Integer totalScore;
-    private Integer passScore;
+    private Double totalScore;
+    private Double passScore;
     private Boolean isPassed;
 
 }
